Fix qemu OCI image build
Temporary fix to make qemu build with our new tkey-app-builder by
installing missing package for Python venv and some more supporting
packages to be able to build.

The 444ee3d26c3acf651ff1bbb12023034ccee6ed68 commit to build the
firmware doesn't seem to exist anymore in tillitis-key1 so we use the
TK1-23.03.2 tag instead.

Of course the digest will be different when we build with new version
of the tools. We have to figure this out somehow, that is build a
Bellatrix version of a more modern codebase.

We probably also want to make the Ubuntu runtime slimmer.